requirements.txt
pyrogram==1.4.16
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
pyrogram
youtube-search-python==1.4.9
yt-dlp==2021.11.10.1
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
pillow
motor
psutil
future
wget
